requirements.txt
gpiozero~=2.0.1
rpi.GPIO
Mock.GPIO
pyyaml
paho-mqtt
pymysql